Show mod file size in mods list
This implements the changes made in PR #48 of the beatmods backend to add file sizes to mods.
The change also allows for warning if a mod is oversized. Most mods shouldn't have a warning, but should a mod go over a limit, it'll warn the user.

THIS WILL REQUIRE FURTHER TESTING ONCE THE BACKEND UPDATES